gpcr pathway sim/gpcr_camp_pka/requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pyyaml>=6.0
streamlit>=1.32
